+++ docs/versionhistory.rst
@@ -3,9 +3,9 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-
 **UNRELEASED**
 
+- Added an asynchronous implementation of the ``functools`` module
 - Set ``None`` as the default type argument for ``anyio.abc.TaskStatus``
 - Added support for ``uvloop=True`` on Windows via the winloop_ implementation
   (`#960 <https://github.com/agronholm/anyio/pull/960>`_; PR by @Vizonex)